Spring Insight service always available
The Spring Insight framework is different than all of the other
services because it downloads its dependency from the service itself,
rather than from a central location.  It does however honor the
Internet availability flag even though it doesn't go to the Internet.
This caused issues when an offline buildpack was used because it
prevented the proper dependency from being downloaded.

This change adds the ability to _temporarily_ override the value of
Internet availability so that the Spring Insight framework can always
download its dependencies, regardless of the current state of internet
connectivity.

[#70850900]